travis: drop XL C/C++ build
xlc works, but it is very slow and uses a ton of memory when compiling
the AVX-512 cmp stuff (or anything which depends on it).  Enabling
optimization (even -O1) helps a lot with the time, but it still gobbles
up too much memory for Travis.

It may be possible to move the build over to GitHub Actions, Azure
Pipelines, etc., though it is also the one compiler we support which
Meson doesn't work with.